% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf2tstat.R
\name{bf2tstat}
\alias{bf2tstat}
\title{Compute t-Statistic from a Bayes Factor}
\usage{
bf2tstat(bayes_factor, sample_size, g = sample_size)
}
\arguments{
\item{bayes_factor}{A numeric value representing the Bayes Factor.}

\item{sample_size}{A numeric value representing the sample size.}

\item{g}{A numeric value representing the prior scaling parameter.
Defaults to \code{g = sample_size}, following Zellner's g-prior.}
}
\value{
A numeric value representing the computed \strong{t-statistic}.
}
\description{
This function calculates the \strong{t-statistic} corresponding to a given Bayes Factor (BF)
and sample size. The method follows the g-prior framework. The transformation used is
derived from eq.5 in Nikolakopoulos, S., & Ntzoufras, I. (2021). Meta Analysis of Bayes Factors.
}
\examples{
# Example: Compute t-statistic for a given Bayes Factor and sample size
bf2tstat(bayes_factor = 5, sample_size = 30)

# Example: Compute t-statistic with a custom g parameter
bf2tstat(bayes_factor = 2.5, sample_size = 30, g = 15)
}
