% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibr.R
\docType{package}
\name{Calibr}
\alias{Calibr}
\alias{Calibr-package}
\title{Calibr: A tool to standardize species population aboundance.}
\description{
Calibr is a R-Package designed to process highly diverse coral reef species datasets and to
analyse with Generalized Linear Models (GLM) to obtian standarization factors.
}
