% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibur_results.R
\name{calibur_results}
\alias{calibur_results}
\title{Returns species specifc data from the reef fish dataset.}
\usage{
calibur_results(SET, std_method)
}
\arguments{
\item{SET}{Coral Reef Dataset}

\item{std_method}{establishes the standard METHOD of the Reef dataset}
}
\description{
Loads data and splits the observaion table
}
