% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_results.R
\name{export_results}
\alias{export_results}
\title{Export results to saved files.}
\usage{
export_results(datalist, outdir = Sys.getenv("HOME"))
}
\arguments{
\item{datalist}{List containing the Grouped dataset and summaries from \code{\link{load_dataset}}}

\item{outdir}{Location to export data and summaries to. Default location is user's \code{HOME} directory}
}
\description{
Option to export grouped survey datasets, the GCF/OPUE Summary Table, and REP summaries in
seperate data files.
}
\details{
These following datalist objects will be written to the location, defined by \code{outdir}.
\itemize{
  \item The Grouped survey dataset object will be as saved as \code{reeffish_grouped.Rds}.
  \item The GCF and OPUE summaries will be saved as \code{summary_table.csv}.
  \item The \code{REP} count by species group summary will be saved as \code{REP_summary.csv}.
}

If the user doesn't specify a output directory, then a new directory will be generated from the user's HOME location.
}
