% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcf.R
\name{gcf}
\alias{gcf}
\title{Gear Calibration Factor}
\usage{
gcf(SET, min_obs = 10)
}
\arguments{
\item{SET}{Coral Reef Species Dataset}

\item{min_obs}{Minimum limit for the number of observations.}
}
\description{
Sets the geat Calibration Factor
}
