% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcf_glm.R
\name{gcf_glm}
\alias{gcf_glm}
\title{Gear Calibration Factor summary table}
\usage{
gcf_glm(SET, min_obs = 10, Standard)
}
\arguments{
\item{SET}{Survey Dataset}

\item{min_obs}{Minimum limit for the number of observations.}
}
\description{
Sets summary table for the Gear Calibration Factor and Observed Effort Per Unit per species group.
}
