% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcf_glm.R
\name{gcf_glm_apply}
\alias{gcf_glm_apply}
\title{Gear Calibration Factor summary table}
\usage{
gcf_glm_apply(SET, std_method, min_obs = 10)
}
\arguments{
\item{SET}{Survey Dataset}

\item{std_method}{Denotes Survey dataset METHOD string as the Standard METHOD}

\item{min_obs}{Minimum limit for the number of observations.}
}
\description{
Sets summary table for the Gear Calibration Factor and Observed Effort Per Unit per species group.
}
