% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcf_glmm.R
\name{gcf_glmm}
\alias{gcf_glmm}
\title{Gear Calibration Factor summary table, hierarchical model version}
\usage{
gcf_glmm(ORIG, min_obs = 10, n_sample = 5, Standard)
}
\arguments{
\item{ORIG}{Original survey dataset}

\item{min_obs}{Minimum limit for the number of observations.}
}
\description{
Sets summary table for the Gear Calibration Factor for all groups analyzed under a single hierarchical, mixed-effect model.
}
