% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reef_datalist.R
\name{get_reef_datalist}
\alias{get_reef_datalist}
\title{Returns species specifc data from the reef fish dataset.}
\usage{
get_reef_datalist(SET, std_method)
}
\arguments{
\item{SET}{Coral Reef Dataset}

\item{std_method}{establishes the standard METHOD of the Reef dataset}
}
\description{
Loads data and splits the observaion table
}
