% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbf2tstat.R
\name{logbf2tstat}
\alias{logbf2tstat}
\title{Compute t-statistic from a log Bayes Factor}
\usage{
logbf2tstat(log_BF, sample_size, g = sample_size)
}
\arguments{
\item{log_BF}{Numeric, log Bayes factor (natural log).}

\item{sample_size}{Numeric, sample size (must be > 2).}

\item{g}{Numeric, prior scaling parameter (default = \code{sample_size}).}
}
\value{
Numeric t-statistic magnitude implied by the Bayes factor.
Returns 0 if the mapping is not real (too small BF).
}
\description{
Inverts the g-prior Bayes factor relationship (Eq. 5) to recover the
corresponding \emph{t}-statistic magnitude. The log Bayes factor is
assumed to be in natural-log units, as returned by \code{tstat2logbf()}.
}
\examples{
t <- 2.5; n <- 30
lb <- tstat2logbf(t, n)
logbf2tstat(lb, n)

}
