% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_summary.R
\name{rep_summary}
\alias{rep_summary}
\title{Site summmary statistics per reef species group}
\usage{
rep_summary(SET, std_method)
}
\arguments{
\item{SET}{Dataset}

\item{std_method}{Establishes the standard METHOD of the Reef dataset}
}
\description{
Calculates descriptive statistics for each GROUP
}
