% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_calibr.R
\name{run_calibr}
\alias{run_calibr}
\title{Survey dataset by species group and summaries}
\usage{
run_calibr(SET, std_method, model_type)
}
\arguments{
\item{SET}{Survey Dataset}

\item{std_method}{Denotes Survey dataset METHOD string as the Standard METHOD}

\item{model_type}{Type of Generalized Linear Model used. TODO RENAME FIELD}
}
\value{
List object with three items:
\itemize{
  \item A list containing the Survey dataset split by species group.
  \item A data.table object with the GCF and OPUE summaries by species group with positive observations.
  \item Data frame summation of number of \code{REP} by species group.
}
}
\description{
Using the input survey dataset, the function does three major tasks. First, it splits
the input dataset by its \code{GROUP} value. The split dataset is used for the
remaining tasks: Summarizing the Gear Calibration Vaules (GCF) and Observed Per Unit
Effort (OPUE) by species group with positive observations, plus giving a summary of
total REP for each species group.
}
