% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SIMDATA}
\alias{SIMDATA}
\title{SIMDATA: Simulation data}
\format{A data.frame with 4586 rows, and 6 fields:
\describe{
 \item{GROUP}{Species or taxinominal grouping.}
 \item{BLOCK}{Spacetime temporial strata.}
 \item{REP}{Replicate, survey site.}
 \item{METHOD}{Survey method used.}
 \item{DENSITY}{Group abundance of positive observations.}
 \item{PRESENCE}{Proportion of presence over absence of a group at a survey site.}
}}
\source{
\url{https://www.pifsc.noaa.gov/library/pubs/admin/PIFSC_Admin_Rep_11-08.pdf}
}
\usage{
SIMDATA
}
\description{
Description Of SIMDATA
}
\keyword{datasets}
