% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SMALL_UNPAIR}
\alias{SMALL_UNPAIR}
\title{NOAA Pacific RAMP, Small Belt Transect, 2002-2010}
\format{A data.frame with 263580 rows, and 6 fields:
\describe{
 \item{BLOCK}{Spacetime temporial strata.}
 \item{REP}{Replicate, survey site.}
 \item{GROUP}{Species or taxinominal grouping.}
 \item{METHOD}{Survey method used.}
 \item{DENSITY}{Group abundance of positive observations.}
 \item{PRESENCE}{Proportion of presence over absence of a group at a survey site.}
}}
\source{
\url{https://www.pifsc.noaa.gov/library/pubs/admin/PIFSC_Admin_Rep_11-08.pdf}
}
\usage{
SMALL_UNPAIR
}
\description{
Dataset from Coral Reef Fish Surveys, conducted by the NOAA Pacific Reef
Assessment and Monitoring Program (RAMP). The survey area includes the
Main and Northwest Hawaiian Islands, the Mariana Archipelago, and various
remote islands in the central pacific.
}
\keyword{datasets}
