% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_dataset.R
\name{calibrate_dataset}
\alias{calibrate_dataset}
\title{Convert original dataset using the calculated GCFs}
\usage{
calibrate_dataset(ORIG, GCFs, Standard)
}
\arguments{
\item{ORIG}{Original Survey Dataset}

\item{GCFs}{Gear calibration factors obtained from gcf_glmm function.}

\item{Standardard}{Standardard gear}
}
\description{
Convert original dataset using the calculated GCFs
}
