% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_dataset.R
\name{calibrate_dataset}
\alias{calibrate_dataset}
\title{Convert original dataset using the calculated GCFs}
\usage{
calibrate_dataset(ORIG, std_method, GCFs)
}
\arguments{
\item{ORIG}{Original Survey Dataset}

\item{std_method}{Standard gear METHOD}

\item{GCFs}{Gear calibration factors obtained from gcf_glmm function.}
}
\description{
Convert original dataset using the calculated GCFs
}
