% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_reef_datalist.R
\name{export_reef_datalist}
\alias{export_reef_datalist}
\title{Export reef data and summaries to data files.}
\usage{
export_reef_datalist(datalist, outdir = Sys.getenv("HOME"))
}
\arguments{
\item{datalist}{Calibr datalist}

\item{outdir}{Location to export data and summaries to. Default location is user's \code{HOME} directory}
}
\description{
Option to export grouped reeffish datasets, Reefish Group Summary Table, and REP summaries in
seperate data files.
}
