% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_reef_dataset.R
\name{load_reef_dataset}
\alias{load_reef_dataset}
\title{Returns the reef survey dataset object from csv file.}
\usage{
load_reef_dataset(file)
}
\arguments{
\item{file}{Reef survey dataet filename}
}
\description{
Loads the Coral Reef Reef Survey dataset from comma seperated values formatted file.
This funcion will throw an error if either the input dataset does not have the NOAA-CRED Pacific-wide
dataset column names ("BLOCK","REP","GROUP","METHOD","DENSITY","PRESENCE"), or it does not not find
exactly 2 unique METHOD values in the input dataset.
}
