% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_reeffish_datalist.R
\name{set_reeffish_datalist}
\alias{set_reeffish_datalist}
\title{Returns species specifc data from the reef fish dataset.}
\usage{
set_reeffish_datalist(SET, std_method)
}
\arguments{
\item{SET}{Coral Reef Dataset}

\item{std_method}{establishes the standard METHOD of the Reef dataset}
}
\description{
Loads data and splits the observaion table
}
